+++ FidelityCard.Application/Services/PromoService.cs
@@ -24,12 +24,10 @@ public class PromoService : IPromoService
 
     public async Task<Guid> Create(PromoRequestDto dto, IFormFile? file)
     {
-        var fileName = file is not null
-            ? await _blobStorage.UploadFile(PromosContainer, file.OpenReadStream(), file.FileName)
-            : null;
-
         var promo = _mapper.Map<Promo>(dto);
-        promo.ImageFileName = fileName;
+        
+        if (file is not null)
+            promo.ImageFileName = await _blobStorage.UploadFile(PromosContainer, file.OpenReadStream(), file.FileName);
 
         _repository.Insert(promo);
         _repository.SaveChanges();
@@ -52,20 +50,21 @@ public class PromoService : IPromoService
 
     public async Task Edit(Guid id, PromoRequestDto dto, IFormFile? file)
     {
-        var promo = _repository.Read(id);
+        var promo = await _repository.GetByIdWithProduct(id);
         if (promo is null)
             throw new ResourceNotFoundException($"Product {id} not found.");
 
-        if (!string.IsNullOrWhiteSpace(promo.ImageFileName))
-            _blobStorage.DeleteFile(PromosContainer, promo.ImageFileName);
+        if (file is not null)
+        {
+            if (!string.IsNullOrWhiteSpace(promo.ImageFileName))
+                _blobStorage.DeleteFile(PromosContainer, promo.ImageFileName);
 
-        var fileName = file is not null
-            ? await _blobStorage.UploadFile(PromosContainer, file.OpenReadStream(), file.FileName)
-            : null;
+            var newFileName = await _blobStorage.UploadFile(PromosContainer, file.OpenReadStream(), file.FileName);
+            promo.ImageFileName = newFileName;
+        }
 
         promo.Name = dto.Name;
         promo.Description = dto.Description;
-        promo.ImageFileName = fileName;
         promo.From = dto.From;
         promo.To = dto.To;
 
@@ -73,18 +72,23 @@ public class PromoService : IPromoService
         _repository.SaveChanges();
     }
 
-    public IEnumerable<PromoResponseDto> GetAll()
+    public IEnumerable<PromoResponseDto> GetByAllWithProduct()
     {
-        foreach (var promo in _repository.List())
+        foreach (var promo in _repository.GetByAllWithProduct())
             yield return _mapper.Map<PromoResponseDto>(promo);
     }
 
-    public PromoResponseDto GetById(Guid id)
+    public async Task<PromoResponseDto> GetByIdWithProduct(Guid id)
     {
-        var promo = _repository.Read(id);
+        var promo = await _repository.GetByIdWithProduct(id);
         if (promo is null)
             throw new ResourceNotFoundException($"Promo {id} not found.");
 
-        return _mapper.Map<PromoResponseDto>(promo);
-    }
+		var promoDto = _mapper.Map<PromoResponseDto>(promo);
+
+		if (!string.IsNullOrWhiteSpace(promo.ImageFileName))
+			promoDto.ImageContent = _blobStorage.DownloadBase64FileContent(PromosContainer, promo.ImageFileName);
+
+		return promoDto;
+	}
 }
